Fix test_triton_kernel_float64_constant
Summary: Landed https://github.com/pytorch/pytorch/pull/135260 too soon and the test in that PR doesn't do exactly what I tested (actually test different dtypes).

Test Plan: `python test/inductor/test_triton_kernels.py -k float64_constant`

[ghstack-poisoned]

+++ test/inductor/test_triton_kernels.py
@@ -1556,7 +1556,7 @@ def forward(self, x_1, output_1):
         def f(x):
             return x * (0.12 * x.shape[0])
 
-        x = torch.ones(200, device=GPU_TYPE, dtype=torch.float64)
+        x = torch.ones(200, device=GPU_TYPE, dtype=dtype)
 
         eager_out = f(x)
         compiled_out = torch.compile(f, dynamic=True)(x)
